# Make some tools dig faster