* **Stone pickaxe:** 1.20 ⇒ 1.00 seconds to dig `cracky = 3` nodes like stone and coal ores. This should make the game feel much better to play.

* **Wooden axe:** 2.00 ⇒ 1.60 seconds to dig `choppy = 3` nodes like trees. Before, the wooden axe was no faster than the hand, as they both had 2.00 seconds digging time.

* **Stone axe:** 1.50 ⇒ 1.30 seconds to dig `choppy = 3` nodes like trees.