@C1ffisme 2015-02-27
I think this pull request can do better. Instead of moving just a few tools, lets move...
Wood to the speed of Stone
Stone to the speed of Iron/Bronze
Iron/Bronze only gets enough of a speed improvement to be worth getting after stone.

See, mining with stone is too 
# Slow
and should really be sped up so you can get to iron fast. Iron is essential for locks on chests, and locked doors. So it should be available fast.