@C1ffisme 2015-02-27
@HybridDog No, the hand should remain where it is. If you need some reason to keep them that fast, just consider that the player can carry 99*27 (2673 blocks! Wow!) oven-sized blocks of pure gold. So technically the player should be able to break stone with creative powers in survival.